fix(security): clamp padding+theme, widen path length, pin submodule (#136)
Four small audit fixes:

M5 — `src/redin/render.odin`: clamp content_rect width/height to 0
before passing to Raylib. A theme with horizontal padding wider than
its host previously produced a negative-extent Rectangle, which
Raylib's scissor and rect-outline routines mishandle (raylib upstream
has had wraparound bugs around negative dimensions).

M6 — `src/redin/parser/theme_parser.odin`: route every f32 → u8
conversion through a single `_to_u8(n: f32) -> u8` helper that
clamps to [0,255]. Without this, out-of-range numbers wrapped mod
256 (`:border-width -1` → 255, `:radius 256.5` → 0), giving a
PUT /aspects body the ability to drive the renderer into weird
states. Regression test in theme_parser_test.odin asserts the clamp
across bg, padding, radius, border-width.

L1 — `src/redin/types/view_tree.odin`: widen `Path.length` from u8
to u16. The field mirrors `len(value)`; combined with the M1 fix
that caps nesting at 256, u16 is comfortably wider than anything
the parser will ever produce. Callers in
view_tree_parser.odin and bridge.odin updated to match.

L3 — `.gitmodules`: pin `branch = main` on the `lib/odin-http`
submodule so a future `git submodule update --remote` follows the
maintainer-intended branch rather than whatever happens to be the
default at the time.

Tests: 134 Fennel, 28 parser (+1 new clamp test), 57 bridge, 27
markdown, 25 input, 4 profile — all green.

+++ src/redin/parser/view_tree_parser.odin
@@ -458,7 +458,7 @@ _flatten :: proc(
 
 	p := make([]u8, len(cur))
 	copy(p, cur[:])
-	append(paths, types.Path{value = p, length = u8(len(p))})
+	append(paths, types.Path{value = p, length = u16(len(p))})
 	append(nodes, n.data)
 	append(parent_indices, parent_idx)
 	append(children_list, types.Children{}) // placeholder
